Fix map
This removes an optimization from one of the 'this' methods
but I was having trouble getting it working again. I think if
we want to have this optimization, we should factor out the
hashtable part of DefaultAssociative into a new type and then
Map can use that type (rather than using DefaultAssociative internals).