Allow asking for per-task runtime-related info from any context.
In certain runtime configurations built with `DEBUG=1`, calling
`chpl_task_getinfoRuntime()` from something that was not a Chapel task
context would cause an assertion failure or internal error.  In all the
cases seen so far, the calls that failed were for legitimate purposes.
Unless we're going to require callers of that function to somehow first
check to see if they're in a task context before making the call, it
needs to work without halting in all cases.  Here, make it return NULL
in both tasking layers when it is called from other than a task.

As part of this, adjust other code so that in all other cases where the
internal checks in question would have caused halts they still do.  Also
while here, adjust comm=ofi so that it no longer assumes that if there
is no `infoRuntime` data for the calling context then this must be the
AM handler.  It might be some other thread instead.