Add `chpl__defaultHash` for `borrowed object?` and adjust class tests for set (#15592)
Add `chpl__defaultHash` for `borrowed object?` and adjust class tests
for set (#15592)

This PR adds an overload of `chpl__defaultHash` for
`borrowed object?`. The overload of `chpl__defaultHash` for
`borrowed object` has been removed, because all non-nilable borrowed
objects will coerce to nilable borrowed object. Futures for sets of
classes that pass after adding the overload have been migrated to
normal tests.

Prior to this PR sets of nilable classes (associative domains of
nilable classes, in truth) failed to instantiate because they were
unable to resolve a call to `chpl__defaultHash`. The compiler can
automatically coerce class type arguments with other management
styles to `borrowed` when resolving a function call. However, the
compiler cannot coerce a nilable class to a non-nilable class, which
is why resolution of `chpl__defaultHash` failed when the set
element type was `borrowed C?`.

---

Testing:

- [x] ALL on linux64 when COMM=none
- [x] ALL on linux64 when COMM=gasnet

---

Reviewed by @mppf.