Let a terminating comm=ofi AM handler out of EAGAIN retries.
Certain libfabric providers, notably RxM, connect endpoints dynamically
and expose delays experienced in doing so by returning EAGAIN from
transaction request calls.  Callers who see this must retry until the
transaction is successful.  It turned out that the comm=ofi code to
handle this sort of thing had a race in it, for a very small program
(hello.chpl) when run with tasks=fifo.  In a 2-node program, we did a
non-blocking executeOn from node 0 to node 1 as part of locale model
setup.  On node 1 we executed the body and then did an AMO back to node
0 to down the endCount.  Comm=ofi does AMOs using AMs, done directly in
the handler, so node 0's AM handler performed the AMO using the CPU and
then sent an 'amDone' response back to node 1.  This 'amDone' PUT ended
up in the EAGAIN retry loop due to RxM dynamic endpoint connection.  In
the meantime, some other task/thread on node 0 which was waiting for the
endCount to drop to zero saw that occur as a result of the AMO.  This
allowed the rest of the program logic to proceed to completion on node
0, whereupon we sent a shutdown message to node 1 and set the flag
telling our own node 0 AM handler to exit.  Node 1 dutifully shut down,
including terminating its AM handler.  Alas, at this point node 0's AM
handler was still getting EAGAIN and retrying its 'amDone' PUT to node
1.  The periodic progress calls made by node 1's AM handler eventually
would have allowed the PUT to complete, but that AM handler had already
terminated.  Node 0's AM handler was thus stuck in the EAGAIN loop
forever.

Here, fix this by adjusting the EAGAIN retry loop to allow AM handlers
to exit early if they see their 'please exit' flag set.

In the particular case where the problem occurred we will thus end up
dropping the 'amDone' indicator on the floor.  That's okay, though,
because the Chapel program semantics only depend on endCount reaching 0,
and we watch for that separately.  The 'amDone' has no Chapel-semantic
value; it just allows the on-stmt wrapper function for the non-blocking
executeOn on node 1 to terminate.