AM-mediated nonfetching AMOs don't need 'done' indicators.
When a nonfetching AMO is done by means of an Active Message, since AMO
AMs are performed by the AM handler itself and we assert send-after-send
and other orderings during initialization, the handler should not need
to send a 'done' indicator after performing the AMO.  If it initiated a
task to perform the AMO then the task would need to do so, because then
the transaction ordering wouldn't apply.  But as long as the handler
does the work itself a 'done' response isn't needed.

Note: the following tests fail on Cray CS with this change, so it seems
the transaction orderings we currently assert are not sufficient to
enforce the Chapel MCM:
release/examples/benchmarks/hpcc/ra-atomics
release/examples/benchmarks/hpcc/variants/ra-unordered-atomics
runtime/configMatters/comm/atomic-stress-multi
runtime/configMatters/comm/unordered/unorderedAtomicsStress