Accept increase in comm for remote domain local DR array
test/optimizations/remoteValueForwarding/serialization/domains.chpl has 1
additional GET for the domain size due to using it in
dsiElementInitializationComplete during an array initialization with a
remote domain. DefaultRectangularArr could cache the number of elements
allocated in its own structure to remove this GET and many others for
cases like this where the domain is remote.