Fix problems with errors for array-of-owned
Errors raised in the chpl__coerceCopy function need to have special treatment
and be reported later in callDestructors.

Removed 'implicit' from 'invalid implicit copy-initialization' in some
errors because the logic in the compiler for it didn't make sense to me
and so it seemed better to leave out that distinction for now.

Added more tests arrays of owned to make sure copy elision can get rid of
some of these errors.

The test arrays/types-resized-array/testOwned.chpl was failing and set me
on this path.