Always use = in this array perf test instead of init
This test has historically measured comms for `=` between
two arrays. After split-init and array initialization work,
it now becomes array initialization, which has 2 more on
statements in this pattern. These on statements are probably
due to the additional call to the postallocator.

This commit adjusts the test to always test assignment.