List _collapse: remove shift arg
_collapse() ignored its shift arg, so remove it so it doesn't mislead
anyone.  None of the callers passed a non-1 value.  If anyone does
want to collapse by more than one, they'll have to implement it, but
won't be surprised by _collapse() ignoring the shift value.

While I was here, have _collapse() try to free memory for the case of
collapsing off the last element of the List, as it does for collapsing
in on an interior element.

That changed the line numbers of an error in List.chpl in a few tests,
so future-proof them with copies of the prediff file from
test/library/standard/List/listAppendLifetime, and change the
List.chpl line numbers in their .good files to nnnn.  (One of the
affected tests has the same name as the source of the prediff file,
but is in a subdirectory.)