Remove local AMO logic from AM-AMO handler.
There isn't any control flow path that would result in the handler for
AM-mediated AMOs dealing with a local target variable, because if an
AMO's variable is local we do the operation on the initiating side of
the comm layer logic.  So, remove the unneeded local-variable logic from
the AMO AM handler.