Enforce nonfetching AMO MCM ordering in bulk rather than individually.
We have been enforcing the MCM-specified ordering for AM-mediated
nonfetching AMOs (nfAMOs) by using blocking AMs, where the initiator
waits for the target side to PUT an 'amDone' indicator back to it after
the AMO has been executed on the target node.  Here, instead enforce the
MCM by using nonblocking AMs (done in an earlier commit), asserting
send-after-send transaction ordering so that those are done in order
with respect to each other, and when a sequence of such nfAMO AMs done
by one task is followed by some other thing with MCM implications
(either a regular RMA or the end of the task), using a blocking no-op AM
to each affected node to make sure that the preceding nfAMO AMs have
completed.  Since AMs are ordered and the blocking no-op AM is done in
the handler just as the earlier nfAMO AMs were, when an initiator sees
that the no-op is done it also knows all the nfAMOs are.

This saves multiple partial network trips for 'amDone' responses in the
AM handlers, since the internal AMs used for nfAMOs are "fast", and
replaces them with a single full network round trip to each affected
node for the no-op AMs.  There is thus a tradeoff as to when it's worth
it.  In this initial implementation, a task switches from using blocking
nfAMO AMs to nonblocking ones after doing just 3 such AMOs.  For sure we
want to do at least the first nfAMO AM in each task as blocking, so that
tasks whose only nfAMO is the downEndcount when they end don't have to
do 2 AMs, the nfAMO and then a no-op for ordering.  The value of 3 was
based on running the many-to-one micro-benchmark on nonfetching AMO ops
with only a single worker thread using the verbs;ofi_rxm provider,
before and after this change.  The latter was about 2x faster, so the
'amDone' responses were taking about half the total nfAMO AM time.

There is one major limitation/wrinkle: we can only do this for tasks
that are on fixed threads with bound tx contexts and thus fixed
libfabric endpoints, because libfabric's transaction ordering only
affects the transactions on a single send/receive endpoint pair.  So if
a task did some nfAMO AMs and then a task yield that switched threads
and thus tx endpoints, a subsequent blocking no-op AM wouldn't ensure
that its previous nonblocking nfAMO AMs had completed.

Related to this, I also added a sanity check which causes an internal
error if a task that thinks it is on a tasking layer "fixed" thread
finds that it has switched threads during a chpl_task_yield() call.  So
as not to cause undue overhead, this check is only enabled with a debug
comm build.

While here, I also restored bitmapSizeof() which was deleted a few PRs
ago because it wasn't needed.  And I fixed a couple of bugs in the
internal bitmap support which hadn't shown up until now because all
previous bitmaps were allocated at the ends of buffered-op descriptors
and here we needed one that stood on its own.