Add FLAG_IGNORE_RUNTIME_TYPE and use it in static type fns
chpl__domainFromArrayRuntimeType calls getArrDomType
which is meant to return the static type of the domain.
But under --baseline, it default-initializes the array in
the process.

See for example this test under --baseline
 arrays/ferguson/array-initialization-patterns/array-initialization-patterns

Add FLAG_IGNORE_RUNTIME_TYPE to allow getArrDomType to opt
in to being treated like type/param fn (and not a runtime type fn).

See also PR #11140 for history.