Add the README and Makefiles for CHPL_LIBFABRIC=libfabric.
Before this we supported only CHPL_LIBFABRIC=system, using pkg-config to
find the includes and libs of a libfabric installation outside of the
Chapel repo proper.  Here, start on real support for using a libfabric
out of our own third-party subdir.

This initial commit is just the README, Makefile, and Makefile.include,
but the Makefiles have been tested in a repo which also contained the
src dir and script changes which will arrive in upcoming commits.

Note that in a departure from previous practice for most other packages,
the Makefile.include here does not define LIBFABRIC_INCLUDE_DIR,
LIBFABRIC_LIB_DIR, or LIBFABRIC_BIN_DIR.  These and the corresponding
Makefile*libfabric-libfabric in runtime/make for the runtime build and
runtime/etc for user program builds are not needed because the libfabric
specific scripts in util/chplenv put the libfabric-related options in
CHPL_THIRD_PARTY_COMPILE_ARGS and CHPL_THIRD_PARTY_LINK_ARGS instead.
I'm inclined to think this is a cleaner solution for third-party compile
and link args, so I'm trying it out here.  It's easy to change in the
future if we decide to switch libfabric to the existing practice or
update the other third-party packages to the new style, either way.