Loosen comm=ofi libfabric version requirements.
We have been enforcing use of libfabric v1.8 for both runtime and user
program builds.  But this was just because when I last updated that it
was the most recent release version in the private libfabric repo we've
been using for CHPL_LIBFABRIC=system.  Here, change to merely requiring
that the libfabric version we link user programs with is the same one
the runtime was built with.

(It is important that these match, because libfabric's struct types can
change members and/or size between minor versions.  Thus linking a user
program against a different libfabric version than the compiled code in
the runtime was built against can lead to mis-referencing libfabric
struct members and/or using struct variables that are less than fully
defined.)