In comm=ofi, do liveness checks.
Dynamic connection establishment as implemented by certain providers
(RxM being one) can put us in a situation where we've reached the open
file limit but we're not getting the corresponding error code (-EMFILE)
from our own calls to libfabric, which would enable us to produce our
specific "open files limit" error message.  Here, implement periodic
liveness checks from node 0 to the other nodes, to determine when this
is so and report it.

The liveness check is just a small AM to the remote node, which is
expected to do an RMA back to node 0 to set a flag.  The logic is coded
so as to minimize the associated overhead, notably by only doing a check
every 10 seconds, and using an adjustable integer counter as a proxy for
the 10-second timeout (since time checks are expensive).

With this commit we get an error exit in the extremeBlock test, which
was the one we've had the most instances of silent hangs in.  However,
notably the error printed is the "open files limit" one, not the "did
not respond to liveness check" message added in this commit.  This seems
to be because the liveness checks add a fi_send() which wasn't there
before, and that returns the -EMFILE that nothing else would have.
Nevertheless, at least now we get an error message instead of a hang.
This may be an indication that it would be sufficient to simply do the
fi_send(), without the AM being required to respond back from the remote
side.  But I'd like to think about that for a bit before adjusting the
code.