Remove where-claused overloads of supportedTypeForBulkTransfer
Remove these overloads for record and tuple types because they relied
on default initializing a sample value, which isn't always possible.
Move their code to the body of the generic overload.

Adjust code in _doDefaultInitSlot for associative domains so that it
avoids default initialization if !isDefaultInitializable, instead of
just isNonNilableClass (the former is true for the latter).

I need to check/adjust the primitive "type has default value" so that
it returns false when the type is a tuple containing a non-default-
initializable element (recursive over elements that are tuples).