Add autoconf-preventative file touching.
I've seen at least one instance in which a libfabric build ran autoconf,
which changed Makefile.in and the configure script from what came out of
the libfabric distribution.  We don't want that to happen; we want to
use third-party sources strictly as-is from our repo.  I haven't been
able to reproduce that autoconf invocation, mainly because I don't have
any way to reproduce the file timestamps that led to it.  Here, adopt
the same solution as we're currently using for hwloc and qthreads, which
is to touch some files so that the Makefile doesn't think it needs to
run autoconf.