Clean up liveness checks.
It looks like we probably don't need to check for responses to liveness
checks.  Just trying to send an AM and checking for unrecoverable errors
is enough.  If the target node cannot be reached we will hear about it.
So remove the "liveness check" specific AM type added in the last commit
and just use nonblocking no-op AMs instead.  Also while here, improve
how we estimate and adjust the interval counter that proxies for our
timer internal.