Adjust the CPU type for Shasta systems.
The introductory product information for the Shasta public announcement
says that the CPUs in the first systems will be "AMD® EPYC® 7002".
These seem to be equivalent to the "rome" codename, and indeed the
default CPU module on one Shasta system I checked is craype-x86-rome.
Here, take this is the least-common-denominator CPU type for Shasta, at
least when we're on a Shasta system and can make use of it.  (When on
other system types, notably for ongoing module builds, we'll still use
sandybridge.)