Ignore chpl_fix_thrown_error/halt in global analysis
The test errhandling/parallel/forall-wrap-first-error shows a case where
code checking a config variable could call deinit on an error, which in
theory could virtually dispatch to an error type later defined, which
used a global variable not yet initialized.

Avoid reporting a compilation error in this case by adding
FLAG_IGNORE_IN_GLOBAL_ANALYSIS and applying it to chpl_fix_thrown_error.
The fix_thrown_error calls should not actually be virtually dispatching
anything in any meaningful way so are safe from the point of view of the
analysis. Additionally avoid problems with halt style error checking by
avoiding the analysis in halt calls. If a halt call is halting the
program due to an error, it does not matter so much if a deinit uses
uninitialized memory.