Switch the domain's list of arrays to a hashtable
Domains keep track of the arrays declared over them so that when a
domain is resized it can resize all the arrays declared over it.
Previously, we used a simple LinkedList to store the arrays, which had
performance issues because removal from a linked list is O(n). For the
most part we got lucky and removal happened in reverse insertion order
so removal was effectively O(1). This was because we did serial init and
deinit of arrays-of-arrays, but there are still cases where the
insertion or removal was parallel, which killed performance. This
includes creating arrays in a task-intent clause (11333) and having a
distributed array of arrays (9414). This significantly improves the
performance of those cases, and does not hurt serial insertion/deletion
performance.

We have long wanted to use a set/hashtable to store a domain's arrays, but
previously we couldn't because using an associative array inside BaseDom and
BaseDist would result in a circular dependency. Now that chpl__hashtable has
been split out we can just use that. This creates a simple set wrapper over
hashtable. We do this instead of using the standard library set to have more
control and to avoid bringing in all of set code to hopefully limit the impact
on compilation speed.

This will also allow us to parallelize array initialization for all
types and array deinitialization in a future PR.

Some technical notes:
 - I had to comment out some `key` printing in chpl__hashtble to avoid
   dynamic dispatch trying to stamp out read/writeThis methods for all
   array types. This resulted in trying to print our arrays of sync and
   we don't support writing syncs.
 - Removing an element that isn't in the set is a noop. This is
   required for optimizations/bulkcomm/bharshbarg/arrayViews, but I made
   inserting an existing element an assertion failure.
 - rootLocaleInitialized is now used earlier, so I moved it to
   ChapelBase.