Warm up the comm layer, in the many-to-one test.
It turns out that comm=ofi can have significant startup overhead which
occurs at the time of the first communication.  Here, add a warm-up
phase to the many-to-one test to get this overhead out of the way before
we start on the measured part of the program.

The need for this is most significant when we're using the RxM utility
provider, which supports RDM endpoints over the verbs and tcp providers.
That provider dynamically connects each endpoint pair at the time the
first transaction is initiated to it.  Many-to-one is a particularly
difficult case, because all the worker endpoints have to be connected to
the single target endpoint.  The measured warm-up time with 14 36-task
worker nodes, when the program is run on 16 Cray CS Broadwell-18 nodes,
is about 4 seconds.  When this time was included in the measured section
it badly affected the reported performance.