Add a new primitive to track referenced modules in a scope and save it in blocks
Basically, handles it on BlockStmts like the useList.  While there, fix a
comment referring to the useList's old name (I think?)

This commit is not sufficient for the tests to work, but with the next commit it
will be - there's some clean up I wanted to do with the next commit so I
separated them.