Refactor set to use chpl__hashtable (#15772)
Refactor set to use chpl__hashtable (#15772)

This PR refactors the set type to make use of `chpl__hashtable`
instead of an associative array. This decouples `set` from any
adjustments that might need to be made to associative arrays in the
future, (i.e. to support non-nilable classes) while letting `set`
rely on a common implementation that is shared with `map` and
associative arrays and domains.

It also adds two tests for parallel zippering (prior to this, it
looks like we had no code testing leader/follower iterators for
set).

Future work:

  - Sets of same length cannot be zippered if their elements
    are distributed poorly and their followers yield differing
    number of full slots: #15824

---

Testing:

- [x] library/standard/Set locally on darwin when quickstart
- [x] ALL on linux64 when COMM=none

---

Reviewed by @mppf. Thanks!