Remove the KNL locale model
Remove the KNL locale model because there is no future Xeon Phi
architecture, there's no real performance benefit over the flat locale
model, and it takes a non-trivial amount of time to build/test the
locale model.

Historically, we have kept the KNL locale model around believing that it
helped us test our locale model interface, but I can't think of a time
where we have benefited from that (I don't have any more or less
confidence in our locale model design choices based on whether KNL
testing is running.)

The KNL locale model was originally an interesting target for us because
the KNL architecture presents different NUMA domains and a separate high
bandwidth memory. While we have the locale model and have functionality
to target HBM with on here.highBandwidthMemory(), this has never been
threaded through to a higher level, which means it's not practical to
allocate on the HBM for a real application. i.e. you can't just create a
block distributed array that's only on the HBM.

Additionally, Xeon Phi is a dead architecture. The KNL architecture is
still interesting to us still because Cori has has a large number of KNL
blades, but the recommended way to run on Cori is in quad-cache mode
(ignore numa aspects of KNL and use the HBM in a cache mode.)

So due to lack of a future for Xeon Phi, only low-level support for HBM,
and no real performance advantage, remove the KNL locale model.

Closes https://github.com/Cray/chapel-private/issues/918