Add non-appearing documentation for CHPL_LIBFABRIC.
Following the lead of CHPL_JEMALLOC, add a documentation block for
CHPL_LIBFABRIC to usingchapel/chplenv, but indent it so it doesn't
appear in the output.  Then, "no doc" the reference in the internal
ChapelEnv module so it doesn't appear either.  But at least now the
info is here if we want it.