Add flag FLAG_LLVM_RETURN_NOALIAS, use it on allocators
The noalias attribute on LLVM functions helps optimizations
to know that the allocated memory can't alias an existing
memory location.

Besides adding the attribute, this change marks _ddata_allocate_noinit
as no longer inline, because if it's inline, the attribute will have
no effect, and I don't see a way to add this attribute to C source
code for clang (so don't see how to add it to an extern function).