Add a comm layer AM "warm-up" perf test.
We recently noticed that for AM-based operations such as executeOn, when
configured with comm=ofi and the verbs+RxM provider and run in 'ml-'
perf testing on a Cray CS system, it takes the many-to-one test a
surprisingly long time (like 3s) to do its very first operation from
each sender.  We know this has something to do with the RxM provider's
dynamic endpoint connection and we suspect there may be some sort of
contention problem with that, since in that perf testing configuration
we end up trying to connect over 500 endpoints all at once.

Here, add a comm layer AM warmup performance test to track this.