Fix some issues in Map using chpl__hashtable
In a couple places map used slot numbers from one table in a different table. Fix
it to use slot numbers from the correct table.

Add 'pragma "unsafe"' to the map init= method to allow it to work with owned
classes.

Make the map.update argument ref or const ref so it can work with owned classes.

Change a variable of type map from const to var in mason. It is being assigned
into a tuple and has an owned class member within the chpl__hashtable.