Test updates from switching from linked list to hashtable
LinkedLists was publicly used before, so update tests that were relying
on that to explicitly `use LinkedLists`. Update printMemAllocs test to
increase the memory printing threshold. The hashtable uses more ram than
the previous threshold so it was adding output we didn't want. Just
increase the allocation size and threshold.