Enable default parsing of files due to nested symbols only
Prior to this change, you could only parse a file by default if there was an
import or use of the top-level module.  If there was only an import or use of a
nested module or symbol in the file, we would ignore that use or import and fail
to parse the file by default, leading to unresolved symbol errors later in
compilation.

I believe it is sufficient to check the first symbol in a nested import or use -
symbols that are available in a file but not defined within it (e.g. due to a
reexport) will still be capable of being found, as a use or import of the scope
where it is actually defined is required.  Either we will find a direct use or
import of the scope where it is defined in the module with the reexport, or
we will find another nested reference that still refers to the original file.