Alter import of Set in the Random module that was working around this issue
The import of Set that was present was originally intended to import the set
type from the Set module, but because we dropped nested imports on the floor, it
was resulting in an error when resolving the module name.  Now that this change
is fixed, we can write the import in its original intended form.