Stop using the TOML library and instead use a fake created internal module
TOML was relying on init_elts_method (or something that relied on it) in
ChapelBase, causing a circular dependency when we tried to import and utilize a
symbol within it, even a simple symbol.  Replace its use with a created module
that lives in the same directory as our overridden ChapelNumLocales library, so
that it can be referenced without worry of additional issues.