Remove _computeNumChunks() in chpl__hashtable that added init comms
_computeNumChunks calls `here`, but `here` is a dummyLocale that lives
on locale 0 until the root locale is initialized. That means that
calling `here` early in setup adds communication. The _computeNumChunks
was used to serialize hashtable init, but it really only did that when
the size was 0/1, but `init_elts_method` will already do that (and it's
careful to use serial init before the root locale is initialized to
avoid comm -- 10541) so we can just remove the call.