Make comm=ofi AM buffer management more resilient to message reordering.
Based on reading the man pages, we expected that the CQ entry that
signaled the provider's release of the AM request multi-receive buffer
after it was used up would always be the last one returned by a given
call to fi_cq_read().  We even had an assertion to that effect.  But
recently that assertion failed -- we received a set of 5 CQ entries
where the first one was a RECV which used up the multi-receive buffer,
the second was the indicator that the buffer had been released, and the
third through fifth were RECVs which had been placed earlier in the
buffer.  My hypothesis is that this happened due to our addition of the
MSG_ORDER_SAS message ordering assertion recently, which ended up
requiring the CQ entries for some messages to be delivered in a
different order than the messages themselves had been dropped into the
buffer.

Discussion about this is ongoing in the libfabric-users mailing list and
apparently there is some belief that what what we saw reflects a bug in
the provider.  (Though if I put myself in a provider developer's shoes,
my own sense is that supporting FI_ORDER_SAS necessarily requires either
having this sort of mismatch between message order in the multi-receive
buffer and message order in the CQ, or moving messages around in the
buffer.  But that doesn't matter for our purposes here.)

In any event, and in the meantime, here make the comm layer less fussy
about all this:
- Don't insist that the buffer use-up CQ event be last, in those read.
- Use 2 half-sized buffers and swap back and forth between them.
- Increase the size of the safety zones at the end of the buffers.