Clean CHPL_ENV_HEADER
This header has apparently caused problems when the set of things put
into it change, and (as I understand it), this was exacerbated by
`make clean` not removing the header.  So this makes `make clean`
do that.

Probably a better way of dealing with this issue would be to have the
Makefile be smarter about when the header needs to be re-generated in
its list of dependences.  Currently, that list is empty which means
that as long as the header exists, we won't regenerate it.  For
example, we could re-generate it whenever printchplenv or any of its
helper scripts change...?