Optimize repeated insertions/deletions from chpl__hashtable
Previously, we would only resize if the number of full slots was half
the capacity, but that still led to a lot of collisions and probing
past deleted entries for repeated insertion/deletion cycles. Change to
resize based on the number of empty slots instead.

This improves the performance of 15874 (domains/elliot/primes):

| LIMIT       | assoc arr (before) | assoc arr (now) | custom HT  |
| ----------- | -----------------: | --------------: | ---------: |
| 1_000_000   |  0.30 s            |  0.08 s         | 0.02 s     |
| 10_000_000  |  4.29 s            |  0.97 s         | 0.21 s     |
| 100_000_000 | 72.33 s            | 15.51 s         | 2.27 s     |

So now increasing the size by 10x, increases execution time by 13x
instead of 16x and raw performance is better.