Remove the single task specialization from parallel iterators
We used to specialize our tuple, range, domain, and array parallel
iterators for the case where only a single task would be created and
instead do a direct for loop to avoid the task creation overhead.
However, this specialization lead to leap cloning, which could
significantly increase the amount of code we're generating. In PR 16383
we optimized single iteration coforalls to avoid task creation, which
eliminates most of the overhead and allows us to remove the single task
specialization here. Note that there is still some overhead because
we're doing end count allocation and atomic manipulation but it's not
nearly as high as it used to be and more importantly we're not spawning
tasks that might get scheduled on different threads and be stuck
waiting so the overhead is pretty small.

Longer term the bulk spawning API should allow us to remove the end
count manipulation, but until then I think the performance overhead for
single task coforalls is worth the code size reduction. I had a really
hard time creating a benchmark that shows this overhead. For a no-op
task spawn the overhead is visible, but I didn't see much overhead for a
spmd+serial-forall version of stream and some other variants.

This reduces the emitted statements for a comm=none Arkouda build by
320K (1.43M -> 1.11M) and should shave ~2 minutes off compilation time.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>