Fix set &= operator
In my haste to implement a temp-less &= operator in #15644, I
implemented -= instead of &=.  I don't believe &= can be implemented
without a temp without iterating over the lhs while modifying it.
Just unconditionally use the version that does use a temp, but
implements the correct operation.

This wasn't caught by the test because the &= test only went through
the temp-full path.  Now that there is only one path again, the
existing test tests it.

For the other operators whose implementation has multiple paths or
isn't symmetric, add additional cases to the tests to hit them.
Those are: &, ^, |