Fix associative domain contains race
Grab the lock when calling contains/dsiMember for associative domains.
This was originally fixed in 2dc2051892, but was reintroduced in 15739
when splitting the hashtable out of DefaultAssociative. This race was
exacerbated by 15918, which caused failures for parSafeMember.chpl.

Fix the race by grabbing the lock, and update parSafeMember to run with
cstdlib atomics too (it was skipped when intrinsics weren't used before
because locks were too slow, but that was back when intrinsics and locks
were the only options.)